get the firts api call

+++ app/src/main/java/com/example/skycast/model/Clouds.kt
@@ -1,5 +1,5 @@
 package com.example.skycast.model
 
 data class Clouds(
-    val all: Long,
+    val all: Int,
 )
