+++ app/src/main/java/com/example/skycast/model/List.kt
@@ -4,13 +4,13 @@ import com.google.gson.annotations.SerializedName
 import kotlin.collections.List
 
 data class List(
-    val dt: Long,
+    val dt: Int,
     val main: Main,
     val weather: List<Weather>,
     val clouds: Clouds,
     val wind: Wind,
-    val visibility: Long,
-    val pop: Long,
+    val visibility: Int,
+    val pop: Float,
     val sys: Sys,
     @SerializedName("dt_txt")
     val dtTxt: String,
