+++ app/src/main/java/com/example/skycast/model/Root.kt
@@ -2,8 +2,8 @@ package com.example.skycast.model
 
 data class Root(
     val cod: String,
-    val message: Long,
-    val cnt: Long,
+    val message: Int,
+    val cnt: Int,
     val list: kotlin.collections.List<List>,
     val city: City,
 )
