+++ app/src/main/java/com/example/skycast/model/Wind.kt
@@ -2,6 +2,6 @@ package com.example.skycast.model
 
 data class Wind(
     val speed: Double,
-    val deg: Long,
+    val deg: Int,
     val gust: Double,
 )
